updating the master branch with current live version
updating master branch with current live version of the wishee site.

+++ wishee_site/public_html/components/com_wishee/views/js/list.js
@@ -3,10 +3,13 @@ jQuery(document).ready(function(){
     var UserGiftListDiv     = jQuery('#user-gift-list');
     
     jQuery(document).on('click', '.delete-gift-btn', function(e) {
-        e.preventDefault();
-        var $giftID = jQuery(e.target).attr("id");
-        jQuery('.loader').addClass('visible');
-        deleteGiftViaAJAX($giftID);
+        var $msg = "Are you sure you want to delete this item?";
+        if (confirm($msg)) {
+            e.preventDefault();
+            var $giftID = jQuery(e.target).attr("id");
+            jQuery('.loader').addClass('visible');
+            deleteGiftViaAJAX($giftID);
+        }
     });
     
     function deleteGiftViaAJAX($giftID) {
@@ -31,7 +34,7 @@ jQuery(document).ready(function(){
                     UserGiftListDiv.append(
                     '<div class="row" style="margin-bottom: 10px; border-bottom: 1px solid #ddd; padding-bottom: 10px;">' +
 
-                        '<div class="small-2 columns">' +
+                        '<div class="col-sm-2">' +
                             '<img ' +
                                 'style="max-height: 200px; width: auto;"' +
                                 'src="' + this.product_image_url + '"' +
@@ -39,23 +42,23 @@ jQuery(document).ready(function(){
                             '/>' +
                         '</div>' +
 
-                        '<div class="small-4 columns">' +
+                        '<div class="col-sm-4">' +
                             this.product_name +
-                            '<p><a style="margin-top: 10px;" class="button hollow" href="' + this.product_store_url + '" target="_blank">' +
+                            '<p><a style="margin-top: 10px;" class="btn btn-default" href="' + this.product_store_url + '" target="_blank">' +
                                 'Buy on Amazon' +
                             '</a></p>' +
                         '</div>' +
 
-                        '<div class="small-1 columns">' +
+                        '<div class="col-sm-1">' +
                             this.product_category +
                         '</div>' +
 
-                        '<div class="small-1 columns" style="text-align: right;">'+
+                        '<div class="col-sm-1" style="text-align: right;">'+
                             '&pound;' + this.product_price.slice(0, -2) + '.' + this.product_price.slice(-2) +
                         '</div>' +
 
-                        '<div class="small-3 columns" style="text-align: right;">' +
-                            '<button type="button" class="button delete-gift-btn" id="' + this.gift_id + '">' +
+                        '<div class="col-sm-3" style="text-align: right;">' +
+                            '<button type="button" class="btn btn-danger delete-gift-btn" id="' + this.gift_id + '">' +
                                 '<i class="fa fa-times"></i> Delete' +
                             '</button>' +
                         '</div>' +
